+++ intervals/parser.py
@@ -46,12 +46,18 @@ class IntervalParser(object):
                 raise IntervalException(str(e))
         return lower, upper, True, True
 
-    def __call__(self, bounds, lower_inc, upper_inc):
+    def __call__(self, bounds, lower_inc=None, upper_inc=None):
         if isinstance(bounds, six.string_types):
-            return self.parse_string(bounds)
+            values = self.parse_string(bounds)
         elif isinstance(bounds, Iterable):
-            return self.parse_sequence(bounds)
+            values = self.parse_sequence(bounds)
         elif hasattr(bounds, 'lower') and hasattr(bounds, 'upper'):
-            return self.parse_object(bounds)
+            values = self.parse_object(bounds)
         else:
-            return self.parse_single_value(bounds)
+            values = self.parse_single_value(bounds)
+        values = list(values)
+        if lower_inc is not None:
+            values[2] = lower_inc
+        if upper_inc is not None:
+            values[3] = upper_inc
+        return values